:rotating_light: fix unquoted port mappings
https://www.jetbrains.com/help/inspectopedia/ComposeUnquotedPorts.html
According to the Compose file specification, mapping ports in the HOST:CONTAINER
format may lead to erroneous results when using a container port lower than 60, because
YAML parses numbers in the format xx:yy as a base-60 value. For this reason, we
recommend always explicitly specifying the port mappings as strings.